Type interval

+++ pkg/tsdb/elasticsearch/client/models.go
@@ -207,7 +207,7 @@ func newAggDef(key string, aggregation *aggContainer) *aggDef {
 
 // HistogramAgg represents a histogram aggregation
 type HistogramAgg struct {
-	Interval    int    `json:"interval,omitempty"`
+	Interval    float64    `json:"interval,omitempty"`
 	Field       string `json:"field"`
 	MinDocCount int    `json:"min_doc_count"`
 	Missing     *int   `json:"missing,omitempty"`
